feat: Actualizando composicion de configuración
